Implements #1 #2 and #3
This commit adds new functions to main:
 - parse_search_query
 - validate_search_query
 - build_content_query
 - build_pagination_url

and updates existing functions;
 - build_query_for_index

The route code is also updated for:
 - index
 - embeddings
 - search_embeddings

It then includes frontend changes to the templates:
 - templates\index.html
 - templates\embeddings.html

and the associated:
 - static\js\views\embeddings.js

All of this works together to enable a new, advanced search functionality. This new search box replaces the previous "Contains" filter UI. The new search box allows the following types of searches:

```contains:"help"&turn:user nocontains:"code"```

For backward compatibility, if no valid query syntax is found in the users' search request, then the contents of the search box is treated like the old 'Contains' filter.

This is tested working with the main index page. I wasn't able to test the correct functionality of the /embeddings page as I couldn't get the UMap embeddings to work (I had to mock out that code with dummy embeddings).